requirements.txt
future~=0.18.2
Pillow~=9.2.0
py2exe~=0.11.1.1